% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betacode.R
\docType{data}
\name{plato_republic}
\alias{plato_republic}
\title{Betacode for Plato's Republic}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 6389 rows and 2 columns.
}
\source{
Vatri, A. and B. McGillivray (2018). "The Diorisis ancient Greek
corpus: linguistics and literature." Research Data Journal for the
Humanities and Social Sciences 3(1): 55-65. Available at
https://figshare.com/articles/dataset/The_Diorisis_Ancient_Greek_Corpus/6187256
}
\usage{
plato_republic
}
\description{
This is the betacode for Plato's Republic, sourced from the \href{https://figshare.com/articles/dataset/The_Diorisis_Ancient_Greek_Corpus/6187256}{Diorisis ancient Greek corpus}.
The dataframe contains two columns: the \code{text} column contains
each sentence in betacode, and the \code{location} column contains the Stephanus
book and page number for the sentence.
}
\keyword{datasets}
