% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betacode.R
\name{unicode_to_betacode}
\alias{unicode_to_betacode}
\title{Unicode to Beta Code Conversion}
\usage{
unicode_to_betacode(unicode_text)
}
\arguments{
\item{unicode_text}{A character vector with Greek Unicode to convert to Beta
Code.}
}
\value{
A character vector with corresponding Beta Code. The function gives a
warning when it encounters characters that cannot be converted to Beta
Code, returning these characters wrapped in brackets.
}
\description{
This function reverses the work of \link{betacode_to_unicode}, converting Ancient
Greek text in Unicode to well-formed Beta Code.
}
\examples{
unicode_text <- betacode_to_unicode(plato_republic$text[1:2])
unicode_text
unicode_to_betacode(unicode_text)

# The function gives a warning when it encounters non-Unicode Greek
# characters, wrapping them in brackets.
unicode_to_betacode(paste(unicode_text, "abc"))
}
