% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/betacode.R
\name{betacode_to_unicode}
\alias{betacode_to_unicode}
\title{Betacode to Unicode conversion}
\usage{
betacode_to_unicode(beta_text)
}
\arguments{
\item{beta_text}{The text to be converted to Unicode}
}
\value{
A string with the Unicode representation of the betacode
}
\description{
Betacode to Unicode conversion
}
\examples{
betacode_to_unicode("*sidw\\\\n e)pi\\\\ qala/tth| po/lis *)assuri/wn")
}
